refactor /common
