+++ trans_rss/version
@@ -1,1 +1,1 @@
-0.5.14
+0.5.15
