+++ trans_rss/web/__init__.py
@@ -1,13 +1,13 @@
 from pywebio.platform.fastapi import webio_routes
 
-from .subscribe import sub_list_page, subscribe_page
-from .logs import log_page
-from .config import config_page
-from .manage import manage_subscribe_page
-from .webhook_type import webhook_type_page
-from .subscribe_type import subscribe_type_page
 from trans_rss.config import config
 
+from .config import config_page
+from .logs import log_page
+from .manage import manage_subscribe_page
+from .subscribe import sub_list_page, subscribe_page
+from .subscribe_type import subscribe_type_page
+from .webhook_type import webhook_type_page
 
 routes = webio_routes(
     {
